requirements.txt
-i https://pypi.org/simple
alembic==1.12.0; python_version >= '3.7'
aniso8601==9.0.1
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1; python_version >= '3.6'
blinker==1.6.3; python_version >= '3.7'
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.3.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
decorator==5.1.1; python_version < '3.11' and python_version >= '3.7'
executing==2.0.0
faker==19.11.0
flask==3.0.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.0.0; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==21.2.0
idna==3.4; python_version >= '3.5'
importlib-metadata==6.8.0; python_version < '3.10'
importlib-resources==6.1.0; python_version < '3.9'
ipdb==0.13.13
ipython==8.12.3; python_version < '3.11' and python_version >= '3.7'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.19.1; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==10.1.0
prompt-toolkit==3.0.39; python_full_version >= '3.7.0'
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.16.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.22
sqlalchemy-serializer==1.4.1
stack-data==0.6.3
tomli==2.0.1; python_version < '3.11' and python_version >= '3.7'
traitlets==5.11.2; python_version >= '3.8'
typing-extensions==4.8.0; python_version < '3.9'
urllib3==2.0.7; python_version >= '3.7'
wcwidth==0.2.8
werkzeug==3.0.0; python_version >= '3.8'
zipp==3.17.0; python_version >= '3.8'
